% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agroeco_zones.R
\docType{data}
\name{agroeco_zones}
\alias{agroeco_zones}
\title{Agro-ecological Zones of Benin}
\format{
A simple feature collection (\code{sf}) with 3 features and 4 fields,
with geometry type \code{MULTIPOLYGON} and projected CRS WGS 84 / UTM zone 31N
(EPSG:32631):
\describe{
\item{OBJECTID}{\code{integer}. Unique identifier for each zone.}
\item{ZONES}{\code{character}. Name of the agro-ecological zone. One of
\code{"Guinean"}, \code{"Sudano Guinean"}, or \code{"Sahelian"}.}
\item{AREA KM2}{\code{double}. Area of the zone in square kilometres}
\item{PERIMETER KM}{\code{double}. Perimeter of the zone in kilometres}
\item{geometry}{\code{sfc_MULTIPOLYGON}. Spatial geometry column
(coordinates in metres, UTM zone 31N).}
}
}
\source{
Stanislas Mahussi Gandaho (2026). \emph{Agro-ecological Zones of Benin}
[Data set]. Distributed as part of the \pkg{<beninr>} R package.
}
\usage{
agroeco_zones
}
\description{
A spatial dataset containing the three agro-ecological zones of Benin,
with their respective areas and perimeters. The zones reflect the
country's climatic and vegetation gradients, ranging from the humid
Guinean zone in the south to the arid Sahelian zone in the north.
}
\examples{
if (requireNamespace("sf")) {

library(sf)

# Loads agro-ecological zone shapefile
data(agroeco_zones)

# Transform to projected system
agroeco_zones <- agroeco_zones \%>\%
  sf::st_transform(crs = "EPSG:32631")
}

}
\keyword{datasets}
