% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{bj_get_country}
\alias{bj_get_country}
\title{Get Benin National Boundary}
\usage{
bj_get_country()
}
\description{
Downloads the national boundary of Benin as an \link[sf:sf]{sf} polygon object..
}
\examples{
\dontrun{
benin <- bj_get_country()
plot(sf::st_geometry(benin))

# Save to GeoPackage
sf::st_write(benin, "benin_boundary.gpkg")
}

}
\seealso{
\code{\link[=bj_get_arcgis_layer]{bj_get_arcgis_layer()}}, \code{\link[=bj_get_departements]{bj_get_departements()}},
\code{\link[=bj_get_arrondissements]{bj_get_arrondissements()}}, \code{\link[=bj_get_villages]{bj_get_villages()}},
\code{\link[=bj_get_zones_sanitaires]{bj_get_zones_sanitaires()}}
}
